Fixed Random Attacks not affecting some enemies
Fixed Random Enemy Attacks not affecting some entities with an AIAttackComponent, for example, Steve, all Tanks, & most enemies with multiple ranged attacks

Removed remove_ground & rain_gold from the rare projectile pool, moved to normal projectile pool.
Reason for this change is the event is rare enough to encounter by default, it's rarer to roll into these attacks and it's even rarer for creatures to use them in a way that influences gameplay.